+++ test/conftest.py
@@ -287,6 +287,7 @@ _LANGUAGE_PYTEST_MARKERS: dict[Language, list[MarkDecorator | Mark]] = {
     Language.ANGULAR: [pytest.mark.angular],
     Language.HTML: [pytest.mark.html],
     Language.SCSS: [pytest.mark.scss],
+    Language.GRAPHQL: [pytest.mark.graphql],
 }